PlutusLedgerApi.QuickCheck.Utils

Documentation

data SizedByteString (n :: Natural) where

Helper for ByteStrings of a fixed length. We don't expose the constructor, instead providing a read-only pattern, as well as an accessor function, to ensure that the size invariant is maintained.

Since: 1.0.0

Bundled Patterns

pattern SizedByteString :: forall (n :: Natural). ByteString -> SizedByteString n
Read-only pattern for accessing the underlying ByteString. Use it just like you would use a data constructor in a pattern match.

Since: 1.0.0

unSizedByteString :: forall (n :: Natural). SizedByteString n -> ByteString

Get the underlying ByteString. It is guaranteed to have the length specified in its type.

Since: 1.0.0

data AsWord64 where

Plutus' ledger API often has to 'fake' Word64 using the much larger Integer type. This helper is designed to generate Integers that fit into Word64.

We don't expose the constructor directly; instead, we provide a read-only pattern and an accessor function.

Since: 1.0.0

Bundled Patterns

pattern AsWord64 :: Integer -> AsWord64
Read-only pattern for accessing the underlying Integer. Use it just like you would use a data constructor in a pattern match.

Since: 1.0.0

fromAsWord64 :: AsWord64 -> Integer

Since: 1.0.0