PlutusLedgerApi.V2.Orphans

Description

QuickCheck orphans (plus a few helpers) for V2 Plutus ledger API types.

Documentation

newtype NonAdaValue

A Value that contains zero Ada.

Note

This is designed to act as a modifier, and thus, we expose the constructor even though it preserves invariants. If you use the constructor directly, be very certain that the Value being wrapped satisfies the invariants described above: failing to do so means all guarantees of this type are off the table.

Since: 1.0.0

Constructors

NonAdaValue Value

getNonAdaValue :: NonAdaValue -> Value

Since: 1.0.0

newtype UTxOValue

A Value suitable for TxOut. Specifically:

The Value is sorted by both keys (meaning CurrencySymbol and TokenName);

There exists an Ada amount; and

All amounts are positive.

Note

This is designed to act as a modifier, and thus, we expose the constructor even though it preserves invariants. If you use the constructor directly, be very certain that the Value being wrapped satisfies the invariants described above: failing to do so means all guarantees of this type are off the table.

Since: 1.0.2

Constructors

UTxOValue Value

getUtxoValue :: UTxOValue -> Value

Since: 1.0.0

newtype FeeValue

A Value containing only Ada, suitable for fees. Furthermore, the Ada quantity is positive.

Note

This is designed to act as a modifier, and thus, we expose the constructor even though it preserves invariants. If you use the constructor directly, be very certain that the Value being wrapped satisfies the invariants described above: failing to do so means all guarantees of this type are off the table.

Since: 1.0.0

Constructors

FeeValue Value

getFeeValue :: FeeValue -> Value

Since: 1.0.0

Orphan instances

Arbitrary Data
This is a very general instance, able to produce Data of basically any shape. You probably want something more focused than this.

Since: 1.0.0

Instance details

Methods

arbitrary :: Gen Data

shrink :: Data -> [Data]

Arbitrary LedgerBytes
Since: 1.0.0

Instance details

Methods

arbitrary :: Gen LedgerBytes

shrink :: LedgerBytes -> [LedgerBytes]

Arbitrary ScriptContext
Since: 1.0.0

Instance details

Methods

arbitrary :: Gen ScriptContext

shrink :: ScriptContext -> [ScriptContext]

CoArbitrary Data
Since: 1.0.0

Instance details

Methods

coarbitrary :: Data -> Gen b -> Gen b

CoArbitrary LedgerBytes
Since: 1.0.0

Instance details

Methods

coarbitrary :: LedgerBytes -> Gen b -> Gen b

CoArbitrary ScriptContext
Since: 1.0.0

Instance details

Methods

coarbitrary :: ScriptContext -> Gen b -> Gen b

Function Data
Since: 1.0.0

Instance details

Methods

function :: (Data -> b) -> Data :-> b

Function LedgerBytes
Since: 1.0.0

Instance details

Methods

function :: (LedgerBytes -> b) -> LedgerBytes :-> b

Function ScriptContext
Since: 1.0.0

Instance details

Methods

function :: (ScriptContext -> b) -> ScriptContext :-> b