PlutusLedgerApi.V3.Orphans

Orphan instances

Arbitrary ChangedParameters
Currently only generates a map with integer keys in the range 0-33, with random values. Does not shrink.

Since: 1.0.1

Instance details

Methods

arbitrary :: Gen ChangedParameters

shrink :: ChangedParameters -> [ChangedParameters]

Arbitrary ColdCommitteeCredential
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen ColdCommitteeCredential

shrink :: ColdCommitteeCredential -> [ColdCommitteeCredential]

Arbitrary Committee
Does not shrink the quorum, as this is surprisingly hard to do sensibly. We assume the quorum is in the interval (0, 1] (meaning anywhere from a single voice to unanimity).

Since: 1.0.1

Instance details

Methods

arbitrary :: Gen Committee

shrink :: Committee -> [Committee]

Arbitrary Constitution
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen Constitution

shrink :: Constitution -> [Constitution]

Arbitrary DRep
This instance has equal chance of generating always-abstain, always-no-confidence and credential 'arms'. Use this instance with this in mind.

Since: 1.0.1

Instance details

Methods

arbitrary :: Gen DRep

shrink :: DRep -> [DRep]

Arbitrary DRepCredential
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen DRepCredential

shrink :: DRepCredential -> [DRepCredential]

Arbitrary Delegatee
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen Delegatee

shrink :: Delegatee -> [Delegatee]

Arbitrary GovernanceAction
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen GovernanceAction

shrink :: GovernanceAction -> [GovernanceAction]

Arbitrary GovernanceActionId
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen GovernanceActionId

shrink :: GovernanceActionId -> [GovernanceActionId]

Arbitrary HotCommitteeCredential
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen HotCommitteeCredential

shrink :: HotCommitteeCredential -> [HotCommitteeCredential]

Arbitrary ProposalProcedure
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen ProposalProcedure

shrink :: ProposalProcedure -> [ProposalProcedure]

Arbitrary ProtocolVersion
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen ProtocolVersion

shrink :: ProtocolVersion -> [ProtocolVersion]

Arbitrary ScriptContext
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen ScriptContext

shrink :: ScriptContext -> [ScriptContext]

Arbitrary ScriptInfo
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen ScriptInfo

shrink :: ScriptInfo -> [ScriptInfo]

Arbitrary ScriptPurpose
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen ScriptPurpose

shrink :: ScriptPurpose -> [ScriptPurpose]

Arbitrary TxCert
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen TxCert

shrink :: TxCert -> [TxCert]

Arbitrary TxInInfo
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen TxInInfo

shrink :: TxInInfo -> [TxInInfo]

Arbitrary TxInfo
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen TxInfo

shrink :: TxInfo -> [TxInfo]

Arbitrary Vote
Does not shrink (as there's not much point).

Since: 1.0.1

Instance details

Methods

arbitrary :: Gen Vote

shrink :: Vote -> [Vote]

Arbitrary Voter
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen Voter

shrink :: Voter -> [Voter]

Arbitrary TxId
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen TxId

shrink :: TxId -> [TxId]

Arbitrary TxOutRef
Since: 1.0.1

Instance details

Methods

arbitrary :: Gen TxOutRef

shrink :: TxOutRef -> [TxOutRef]

CoArbitrary ChangedParameters
Since: 1.0.1

Instance details

Methods

coarbitrary :: ChangedParameters -> Gen b -> Gen b

CoArbitrary ColdCommitteeCredential
Since: 1.0.1

Instance details

Methods

coarbitrary :: ColdCommitteeCredential -> Gen b -> Gen b

CoArbitrary Committee
Since: 1.0.1

Instance details

Methods

coarbitrary :: Committee -> Gen b -> Gen b

CoArbitrary Constitution
Since: 1.0.1

Instance details

Methods

coarbitrary :: Constitution -> Gen b -> Gen b

CoArbitrary DRep
since 1.0.0

Instance details

Methods

coarbitrary :: DRep -> Gen b -> Gen b

CoArbitrary DRepCredential
Since: 1.0.1

Instance details

Methods

coarbitrary :: DRepCredential -> Gen b -> Gen b

CoArbitrary Delegatee
Since: 1.0.1

Instance details

Methods

coarbitrary :: Delegatee -> Gen b -> Gen b

CoArbitrary GovernanceAction
Since: 1.0.1

Instance details

Methods

coarbitrary :: GovernanceAction -> Gen b -> Gen b

CoArbitrary GovernanceActionId
Since: 1.0.1

Instance details

Methods

coarbitrary :: GovernanceActionId -> Gen b -> Gen b

CoArbitrary HotCommitteeCredential
Since: 1.0.1

Instance details

Methods

coarbitrary :: HotCommitteeCredential -> Gen b -> Gen b

CoArbitrary ProposalProcedure
Since: 1.0.1

Instance details

Methods

coarbitrary :: ProposalProcedure -> Gen b -> Gen b

CoArbitrary ProtocolVersion
Since: 1.0.1

Instance details

Methods

coarbitrary :: ProtocolVersion -> Gen b -> Gen b

CoArbitrary ScriptPurpose
Since: 1.0.1

Instance details

Methods

coarbitrary :: ScriptPurpose -> Gen b -> Gen b

CoArbitrary TxCert
Since: 1.0.1

Instance details

Methods

coarbitrary :: TxCert -> Gen b -> Gen b

CoArbitrary Vote
Since: 1.0.1

Instance details

Methods

coarbitrary :: Vote -> Gen b -> Gen b

CoArbitrary Voter
Since: 1.0.1

Instance details

Methods

coarbitrary :: Voter -> Gen b -> Gen b

CoArbitrary TxId
Since: 1.0.1

Instance details

Methods

coarbitrary :: TxId -> Gen b -> Gen b

CoArbitrary TxOutRef
Since: 1.0.1

Instance details

Methods

coarbitrary :: TxOutRef -> Gen b -> Gen b

Function ChangedParameters
Since: 1.0.1

Instance details

Methods

function :: (ChangedParameters -> b) -> ChangedParameters :-> b

Function ColdCommitteeCredential
Since: 1.0.1

Instance details

Methods

function :: (ColdCommitteeCredential -> b) -> ColdCommitteeCredential :-> b

Function Committee
Since: 1.0.1

Instance details

Methods

function :: (Committee -> b) -> Committee :-> b

Function Constitution
Since: 1.0.1

Instance details

Methods

function :: (Constitution -> b) -> Constitution :-> b

Function DRep
Since: 1.0.1

Instance details

Methods

function :: (DRep -> b) -> DRep :-> b

Function DRepCredential
Since: 1.0.1

Instance details

Methods

function :: (DRepCredential -> b) -> DRepCredential :-> b

Function Delegatee
Since: 1.0.1

Instance details

Methods

function :: (Delegatee -> b) -> Delegatee :-> b

Function GovernanceAction
Since: 1.0.1

Instance details

Methods

function :: (GovernanceAction -> b) -> GovernanceAction :-> b

Function GovernanceActionId
Since: 1.0.1

Instance details

Methods

function :: (GovernanceActionId -> b) -> GovernanceActionId :-> b

Function HotCommitteeCredential
Since: 1.0.1

Instance details

Methods

function :: (HotCommitteeCredential -> b) -> HotCommitteeCredential :-> b

Function ProposalProcedure
Since: 1.0.1

Instance details

Methods

function :: (ProposalProcedure -> b) -> ProposalProcedure :-> b

Function ProtocolVersion
Since: 1.0.1

Instance details

Methods

function :: (ProtocolVersion -> b) -> ProtocolVersion :-> b

Function ScriptPurpose
Since: 1.0.1

Instance details

Methods

function :: (ScriptPurpose -> b) -> ScriptPurpose :-> b

Function TxCert
Since: 1.0.1

Instance details

Methods

function :: (TxCert -> b) -> TxCert :-> b

Function Vote
Since: 1.0.1

Instance details

Methods

function :: (Vote -> b) -> Vote :-> b

Function Voter
Since: 1.0.1

Instance details

Methods

function :: (Voter -> b) -> Voter :-> b

Function TxId
Since: 1.0.1

Instance details

Methods

function :: (TxId -> b) -> TxId :-> b

Function TxOutRef
Since: 1.0.1

Instance details

Methods

function :: (TxOutRef -> b) -> TxOutRef :-> b